backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.1.0
langchain-community==0.0.13
pinecone==1.0.0
pypdf==3.17.1
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
streamlit==1.29.0 
pymilvus==2.5.3
chromadb
